almsot done
